Gui
I finally got the Gui to work. It now pops up with the default
background when I right click the Enhancement Table. Now I need to make
a container, a (real) gui and a tile entity associated with the
Enhancement Table. The main problem I ran into was I wasnt instantiating
the GuiHandler, meaning it didn't actually know what to do causing a
crash.

Separated Gui handling from the Common Proxy to it's own class.